requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==3.2.20
django-activity-stream==1.4.2
django-crispy-forms==2.0
django-forms-dynamic==1.0.0
django-htmx==1.15.0
django-tables2==2.6.0
django-widget-tweaks==1.4.12
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
